+++ src/main/java/pages/CodeTabPage.java
@@ -3,8 +3,6 @@ package pages;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.Wait;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 
@@ -25,7 +23,7 @@ public class CodeTabPage extends BasePage {
     public CodeTabPage verifyRepositoryName(String name) {
         WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(2));
         wait.until(ExpectedConditions.visibilityOfElementLocated(repositoryNameLocator));
-        Assertions.assertEquals(driver.findElement(repositoryNameLocator).getText(), name);
+        Assert.assertEquals(driver.findElement(repositoryNameLocator).getText(), name);
         return this;
     }
