+++ src/test/java/api/BaseApiTest.java
@@ -6,13 +6,13 @@ import io.restassured.filter.log.LogDetail;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
 import io.restassured.specification.ResponseSpecification;
-import org.junit.jupiter.api.BeforeEach;
+import org.testng.annotations.BeforeMethod;
 
 public class BaseApiTest {
     protected RequestSpecification requestSpecification;
     protected ResponseSpecification responseSpecification;
 
-    @BeforeEach
+    @BeforeMethod
     public void setSpecs() {
         requestSpecification = new RequestSpecBuilder()
                 .log(LogDetail.ALL)
