+++ src/test/java/ui/BaseTestClass.java
@@ -2,8 +2,6 @@ package ui;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
@@ -12,7 +10,6 @@ import org.testng.annotations.BeforeMethod;
 
 import java.util.concurrent.TimeUnit;
 
-import static helpers.Level.DEBUG;
 import static helpers.Level.INFO;
 import static helpers.PrinterColors.printColorMessage;
