+++ src/test/java/ui/GitHubTest.java
@@ -1,9 +1,13 @@
 package ui;
 
+import org.openqa.selenium.interactions.Actions;
+import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import pages.HomePage;
 import pages.MainPage;
+
+import java.util.List;
 
 import static org.testng.Assert.assertTrue;
 
@@ -29,6 +33,22 @@ public class GitHubTest extends BaseTestClass {
         HomePage homePage = new HomePage(driver);
         homePage.goToLoginPage().negativeLogin("123", "123")
                 .validateErrorMessage("Incorrect username or password.");
+    }
+
+    @Test
+    public void verifyRepositoriesList() {
+        HomePage homePage = new HomePage(driver);
+        List<String> actualRepositories = homePage.goToLoginPage().validateAuthFieldsAreDisplayed()
+                .loginSuccessful("nataliia.rudenko2012@gmail.com", "nataliia.rudenko2012")
+                .goToProfileForm()
+                .goToRepositoriesPage()
+                .getRepositories();
+        List<String> expectedRepositories = List.of("courseIteaFramework", "courseITEA", "myITEAFramework", "TestRepository");
+        Assert.assertEquals(actualRepositories, expectedRepositories, "ActualRepositories not matched " +
+                "with ExpectedRepositories");
+
+        Actions actions = new Actions(driver);
+
     }
 
     @DataProvider(name = "testDataProvider")
